+++ crates/ruma-common/CHANGELOG.md
@@ -9,6 +9,7 @@ Improvements:
 - Add `M_CONCURRENT_WRITE` error code, used by [MSC4438].
 - Add `canonical_json::RedactingSerializer` to serialize a `CanonicalJsonObject`
   while redacting it on the fly.
+- Add `StatusProfileField::new()` and `CallProfileField::new()` constructors for MSC4426
 
 [MSC4438]: https://github.com/matrix-org/matrix-spec-proposals/pull/4438